Prevent wrong builds from not tracking which state of source file target depends on
Keep past states of a source in the db along with current one and made
dependencies db table track which actual state it depended on. Not keeping
track of this can lead to wrong builds because current target could have
been skipped in the last build yet source node past state updated during
update of some other target depending on it too thus making it appear
unchanged this run.