Implement customizable deciders that work per-task
Used this to use pure timestamp method to check for the need to remake
the makefile. This will make it properly up-to-date in case when it was
just built by outer build tool such as qmake. Default database decider
would always rebuild it since all sources are new on first run and new
source always causes a rebuild.